Struct libra_store::db::ColumnFamilyOptions

pub struct ColumnFamilyOptions {
    pub inner: *mut Options,
    pub titan_inner: *mut DBTitanDBOptions,
    // some fields omitted
}

Fields
inner: *mut Optionstitan_inner: *mut DBTitanDBOptions
Methods

impl ColumnFamilyOptions

pub fn new() -> ColumnFamilyOptions

pub unsafe fn from_raw(inner: *mut Options) -> ColumnFamilyOptions

pub fn set_titandb_options(&mut self, opts: &TitanDBOptions)

pub fn optimize_level_style_compaction(&mut self, memtable_memory_budget: i32)

pub fn set_env(&mut self, env: Arc<Env>)

pub fn set_compaction_filter<S>(
&mut self,
name: S,
ignore_snapshots: bool,
filter: Box<dyn CompactionFilter + 'static>
) -> Result<(), String> where
S: Into<Vec<u8>>,

Set compaction filter.

filter will be dropped when this option is dropped or a new filter is set.

By default, compaction will only pass keys written after the most recent call to GetSnapshot() to filter. However, if ignore_snapshots is set to true, even if the keys were written before the last snapshot will be passed to filter too. For more details please checkout rocksdb's documentation.

See also CompactionFilter.

pub fn add_table_properties_collector_factory(
&mut self,
fname: &str,
factory: Box<dyn TablePropertiesCollectorFactory + 'static>
)

pub fn compression(&mut self, t: DBCompressionType)

pub fn get_compression(&self) -> DBCompressionType

pub fn compression_per_level(&mut self, level_types: &[DBCompressionType])

pub fn get_compression_per_level(&self) -> Vec<DBCompressionType>

pub fn bottommost_compression(&self, c: DBCompressionType)

pub fn add_merge_operator(
&mut self,
name: &str,
merge_fn: fn(&[u8], Option<&[u8]>, &mut MergeOperands) -> Vec<u8>
)

pub fn add_comparator(
&mut self,
name: &str,
compare_fn: fn(&[u8], &[u8]) -> i32
)

pub fn set_block_cache_size_mb(&mut self, cache_size: u64)

pub fn set_min_write_buffer_number(&mut self, nbuf: i32)

pub fn set_max_write_buffer_number(&mut self, nbuf: i32)

pub fn set_write_buffer_size(&mut self, size: u64)

pub fn set_max_bytes_for_level_base(&mut self, size: u64)

pub fn set_max_bytes_for_level_multiplier(&mut self, mul: i32)

pub fn get_max_bytes_for_level_multiplier(&mut self) -> i32

pub fn set_max_compaction_bytes(&mut self, bytes: u64)

pub fn set_level_compaction_dynamic_level_bytes(&mut self, v: bool)

pub fn get_level_compaction_dynamic_level_bytes(&self) -> bool

pub fn set_soft_pending_compaction_bytes_limit(&mut self, size: u64)

pub fn get_soft_pending_compaction_bytes_limit(&self) -> u64

pub fn set_hard_pending_compaction_bytes_limit(&mut self, size: u64)

pub fn get_hard_pending_compaction_bytes_limit(&self) -> u64

pub fn set_target_file_size_base(&mut self, size: u64)

pub fn get_target_file_size_base(&self) -> u64

pub fn set_min_write_buffer_number_to_merge(&mut self, to_merge: i32)

pub fn set_level_zero_file_num_compaction_trigger(&mut self, n: i32)

pub fn set_level_zero_slowdown_writes_trigger(&mut self, n: i32)

pub fn get_level_zero_slowdown_writes_trigger(&self) -> u32

pub fn set_level_zero_stop_writes_trigger(&mut self, n: i32)

pub fn get_level_zero_stop_writes_trigger(&self) -> u32

pub fn set_compaction_style(&mut self, style: DBCompactionStyle)

pub fn compaction_priority(&mut self, priority: CompactionPriority)

pub fn set_disable_auto_compactions(&mut self, disable: bool)

pub fn get_disable_auto_compactions(&self) -> bool

pub fn set_block_based_table_factory(&mut self, factory: &BlockBasedOptions)

pub fn set_report_bg_io_stats(&mut self, enable: bool)

pub fn set_num_levels(&mut self, n: i32)

pub fn get_num_levels(&self) -> usize

pub fn set_prefix_extractor<S>(
&mut self,
name: S,
transform: Box<dyn SliceTransform + 'static>
) -> Result<(), String> where
S: Into<Vec<u8>>,

pub fn set_optimize_filters_for_hits(&mut self, v: bool)

pub fn set_memtable_insert_hint_prefix_extractor<S>(
&mut self,
name: S,
transform: Box<dyn SliceTransform + 'static>
) -> Result<(), String> where
S: Into<Vec<u8>>,

pub fn set_memtable_prefix_bloom_size_ratio(&mut self, ratio: f64)

pub fn get_block_cache_usage(&self) -> u64

pub fn set_block_cache_capacity(&self, capacity: u64) -> Result<(), String>

pub fn get_block_cache_capacity(&self) -> u64

pub fn set_fifo_compaction_options(&mut self, fifo_opts: FifoCompactionOptions)

pub fn set_vector_memtable_factory(&mut self, reserved_bytes: u64)