Trait libra_store::db::DBView

pub trait DBView {
type DBSchema: Schema;
    fn get_schema(
        &self, 
        schema: &Self::DBSchema
    ) -> Result<Option<Self::DBSchema>>;

    fn put_schema(&self, schema: Self::DBSchema) -> Result<()>;

    fn write_schemas(&self, batch: SchemaBatch<Self::DBSchema>) -> Result<()>;

    fn write_schemas_index(
        &self, 
        batch: SchemaBatchWithIndex<Self::DBSchema>
    ) -> Result<()>;
}

DBView is an abstraction that defines the interfaces for both DB and ScratchPad. Code works on top of storage should work on DBView, regardless of whether the underlying storage is a DB or a ScratchPad.

Associated Types

type DBSchema: Schema

The schema of the underlying storage.

Loading content...
Required methods

fn get_schema(&self, schema: &Self::DBSchema) -> Result<Option<Self::DBSchema>>

Reads schema from storage.

fn put_schema(&self, schema: Self::DBSchema) -> Result<()>

Writes schema to storage.

fn write_schemas(&self, batch: SchemaBatch<Self::DBSchema>) -> Result<()>

Writes batch to storage.

fn write_schemas_index(
&self,
batch: SchemaBatchWithIndex<Self::DBSchema>
) -> Result<()>

Writes batch to storage.

Loading content...